requirements.txt
# STCI - Standard Token Cost Index
# Python dependencies

# Core
requests>=2.28.0
pyyaml>=6.0

# Validation
jsonschema>=4.17.0

# API (optional - for services/api)
# fastapi>=0.100.0
# uvicorn>=0.22.0

# Development / Testing
pytest>=7.0.0
pytest-cov>=4.0.0
responses>=0.23.0
